+++ src/server/sqlr-listener.cpp
@@ -39,7 +39,7 @@ static void shutDown(int32_t signum) {
 		filename.append(backtrace);
 		filename.append(sys::getDirectorySeparator());
 		filename.append("sqlr-listener.");
-		filename.append(process::getProcessId());
+		filename.append((uint32_t)process::getProcessId());
 		filename.append(".bt");
 		process::backtrace(filename.getString());
 	}
